# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.FavoritesResponse:
    description: Ответ со списком избранных материалов
    properties:
      materials:
        description: Замените на конкретный тип когда будет реализовано
        items: {}
        type: array
      total:
        example: 5
        type: integer
    type: object
...
  handlers.InternalErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Internal server error
        type: string
    type: object
...
  handlers.InvalidParametersErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.MarkCompleteRequest:
    description: Запрос на отметку материала как завершенного
    properties:
      grade:
        example: 4.5
        maximum: 5
        minimum: 1
        type: number
      timeSpent:
        example: 3600
        type: integer
    required:
    - grade
    - timeSpent
    type: object
  handlers.MarkCompleteResponse:
    description: Ответ на отметку материала как завершенного
    properties:
      materialID:
        example: 1
        type: integer
      message:
        example: Material marked as completed
        type: string
    type: object
...
  handlers.ToggleFavoriteRequest:
    description: Запрос на добавление/удаление из избранного
    properties:
      action:
        enum:
        - add
        - remove
        example: add
        type: string
    required:
    - action
    type: object
  handlers.ToggleFavoriteResponse:
    description: Ответ на добавление/удаление из избранного
    properties:
      action:
        example: add
        type: string
      materialID:
        example: 1
        type: integer
      message:
        example: Favorite updated successfully
        type: string
    type: object
...
  models.ProgressMaterial:
    description: Материал с информацией о прогрессе
    properties:
      id:
        example: 1
        type: integer
      lastActivity:
        example: "2023-01-15T10:30:00Z"
        type: string
      progress:
        description: 0-100%
        example: 75.5
        type: number
      subject:
        example: math
        type: string
      title:
        example: Основы алгебры
        type: string
    type: object
...
  models.StudentProgress:
    description: Прогресс обучения студента
    properties:
      averageGrade:
        example: 4.2
        type: number
      completedTopics:
        example: 15
        type: integer
      currentMaterials:
        items:
          $ref: '#/definitions/models.ProgressMaterial'
        type: array
      learningHours:
        example: 120
        type: integer
      successRate:
        example: 85.5
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для образовательной платформы Пайдея
  title: Paydeya Education Platform API
  version: "1.0"
paths:
...
  /student/favorites:
    get:
      consumes:
      - application/json
      description: Возвращает список избранных материалов пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список избранных материалов
          schema:
            $ref: '#/definitions/handlers.FavoritesResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить избранные материалы
      tags:
      - progress
  /student/materials/{id}/complete:
    post:
      consumes:
      - application/json
      description: Отмечает материал как завершенный с оценкой и временем изучения
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Данные завершения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.MarkCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Материал отмечен как завершенный
          schema:
            $ref: '#/definitions/handlers.MarkCompleteResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Отметить материал как завершенный
      tags:
      - progress
  /student/materials/{id}/favorite:
    post:
      consumes:
      - application/json
      description: Добавляет или удаляет материал из избранного
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Действие с избранным
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ToggleFavoriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Статус избранного обновлен
          schema:
            $ref: '#/definitions/handlers.ToggleFavoriteResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавить/удалить из избранного
      tags:
      - progress
  /student/progress:
    get:
      consumes:
      - application/json
      description: Возвращает прогресс обучения текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Прогресс обучения
          schema:
            $ref: '#/definitions/models.StudentProgress'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить прогресс обучения
      tags:
      - progress
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...